frastgl

+++ frast2/flat/main.cc
@@ -5,16 +5,28 @@ using namespace frast;
 
 int main() {
 
-	std::string fname = "flat.it";
-	unlink(fname.c_str());
 
+	std::string fname = "/data/naip/mocoNaip/moco.fft";
+	unlink(fname.c_str());
 	ConvertConfig cfg;
 	cfg.srcPaths = {"/data/naip/mocoNaip/whole.wm.tif"};
-	cfg.baseLevel = 18;
+	cfg.baseLevel = 15;
 	cfg.addo = true;
+	bool isTerrain = false;
+
+	if (0) {
+	fname = "/data/elevation/gmted/gmted.fft";
+	unlink(fname.c_str());
+	cfg.srcPaths = {"/data/elevation/gmted/usa_mean150.3857_4979.tif"};
+	cfg.baseLevel = 8;
+	cfg.addo = true;
+	isTerrain = true;
+	cfg.channels = 1;
+	}
 
 	{
 		EnvOptions envOpts;
+		envOpts.isTerrain = isTerrain;
 		WriterMaster wm(fname, envOpts);
 		wm.start(cfg);
 
@@ -26,6 +38,7 @@ int main() {
 
 	if (cfg.addo) {
 		EnvOptions envOpts;
+		envOpts.isTerrain = isTerrain;
 		WriterMasterAddo wm(fname, envOpts);
 		wm.start(cfg);
